Prefix nginx configuration files to get it parsed before anything
I had funny failling cases : app confs that were in lexicographic order before
"letsencrypt". The ACME challenge was not reachable for those.

That does not hurt all applications, but those who are at the root of the
domain (not in a subfolder) are impacted.

That patch gives better chances that the letsencrypt conf has priority over the
application itself, making acme challenge reachable.